Move the power-related functions to util
