+++ plugwise/common.py
@@ -15,43 +15,16 @@ from plugwise.constants import (
     ApplianceType,
     GwEntityData,
     ModuleData,
-    SensorType,
 )
 from plugwise.util import (
-    check_alternative_location,
     check_heater_central,
     check_model,
     get_vendor_name,
-    power_data_local_format,
     return_valid,
 )
 
 from defusedxml import ElementTree as etree
 from munch import Munch
-
-
-def collect_power_values(
-    data: GwEntityData, loc: Munch, tariff: str, legacy: bool = False
-) -> None:
-    """Something."""
-    for loc.peak_select in ("nl_peak", "nl_offpeak"):
-        loc.locator = (
-            f'./{loc.log_type}[type="{loc.measurement}"]/period/'
-            f'measurement[@{tariff}="{loc.peak_select}"]'
-        )
-        if legacy:
-            loc.locator = (
-                f"./{loc.meas_list[0]}_{loc.log_type}/measurement"
-                f'[@directionality="{loc.meas_list[1]}"][@{tariff}="{loc.peak_select}"]'
-            )
-
-        loc = power_data_peak_value(loc, legacy)
-        if not loc.found:
-            continue
-
-        data = power_data_energy_diff(loc.measurement, loc.net_string, loc.f_val, data)
-        key = cast(SensorType, loc.key_string)
-        data["sensors"][key] = loc.f_val
 
 
 def get_zigbee_data(module: etree, module_data: ModuleData, legacy: bool) -> None:
@@ -67,62 +40,6 @@ def get_zigbee_data(module: etree, module_data: ModuleData, legacy: bool) -> Non
     elif (zb_node := module.find("./protocols/zig_bee_node")) is not None:
         module_data["zigbee_mac_address"] = zb_node.find("mac_address").text
         module_data["reachable"] = zb_node.find("reachable").text == "true"
-
-
-def power_data_peak_value(loc: Munch, legacy: bool) -> Munch:
-    """Helper-function for _power_data_from_location() and _power_data_from_modules()."""
-    loc.found = True
-    if loc.logs.find(loc.locator) is None:
-        loc = check_alternative_location(loc, legacy)
-        if not loc.found:
-            return loc
-
-    if (peak := loc.peak_select.split("_")[1]) == "offpeak":
-        peak = "off_peak"
-    log_found = loc.log_type.split("_")[0]
-    loc.key_string = f"{loc.measurement}_{peak}_{log_found}"
-    if "gas" in loc.measurement or loc.log_type == "point_meter":
-        loc.key_string = f"{loc.measurement}_{log_found}"
-    # Only for P1 Actual -------------------#
-    if "phase" in loc.measurement:
-        loc.key_string = f"{loc.measurement}"
-    # --------------------------------------#
-    loc.net_string = f"net_electricity_{log_found}"
-    val = loc.logs.find(loc.locator).text
-    loc.f_val = power_data_local_format(loc.attrs, loc.key_string, val)
-
-    return loc
-
-
-def power_data_energy_diff(
-    measurement: str,
-    net_string: SensorType,
-    f_val: float | int,
-    data: GwEntityData,
-) -> GwEntityData:
-    """Calculate differential energy."""
-    if (
-        "electricity" in measurement
-        and "phase" not in measurement
-        and "interval" not in net_string
-    ):
-        diff = 1
-        if "produced" in measurement:
-            diff = -1
-        if net_string not in data["sensors"]:
-            tmp_val: float | int = 0
-        else:
-            tmp_val = data["sensors"][net_string]
-
-        if isinstance(f_val, int):
-            tmp_val += f_val * diff
-        else:
-            tmp_val += float(f_val * diff)
-            tmp_val = float(f"{round(tmp_val, 3):.3f}")
-
-        data["sensors"][net_string] = tmp_val
-
-    return data
 
 
 class SmileCommon:
