+++ plugwise/helper.py
@@ -8,7 +8,7 @@ from __future__ import annotations
 import datetime as dt
 from typing import cast
 
-from plugwise.common import SmileCommon, collect_power_values
+from plugwise.common import SmileCommon
 from plugwise.constants import (
     ACTIVE_ACTUATORS,
     ACTUATOR_CLASSES,
@@ -44,6 +44,7 @@ from plugwise.constants import (
 )
 from plugwise.util import (
     check_model,
+    collect_power_values,
     common_match_cases,
     count_data_items,
     format_measure,
